Listen for "stream paused" events
Listen for the stream state change event that gets emitted when Pipewire
pauses the capture links, and generate a new internal event. This is
then used to trigger the "off" sound (if needed) and hide the tray icon.

Playing the clicks has been moved to its own thread. This simplifies the
event dispatch logic, as events can now be generated in two different
callback functions.